refactor(agents): move token format helpers to the agents feature root
formatTokenCompact and formatTokenExact serve the model picker, catalog
select options, and the context usage indicator, so agent-chat is not
their home. Move them up one level and update the three importers.
